# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GlobalLand API
  description: |
    The GlobalLand API provides comprehensive endpoints for real estate tokenization, 
    investment management, and trading on the Hedera blockchain. This RESTful API 
    enables developers to build applications that interact with tokenized real estate assets.
  version: 1.0.0
  contact:
    name: GlobalLand Support
    email: support@globalland.com
    url: https://docs.globalland.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.globalland.com/api
    description: Production server
  - url: https://sandbox-api.globalland.com/api
    description: Sandbox server
security:
  - BearerAuth: []
paths:
...
  /properties:
    get:
      tags:
        - Properties
      summary: Get all properties
      description: Retrieve list of available properties with filtering and pagination
      security: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: location
          in: query
          description: Filter by location
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price per token
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price per token
          schema:
            type: number
            minimum: 0
        - name: propertyType
          in: query
          description: Property type filter
          schema:
            $ref: '#/components/schemas/PropertyType'
      responses:
        '200':
          description: Properties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesResponse'

    post:
      tags:
        - Properties
      summary: Create new property
      description: Register a new property for tokenization (Property Manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreation'
      responses:
        '201':
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PropertyType:
      type: string
      enum: [RESIDENTIAL, COMMERCIAL, INDUSTRIAL, LAND, MIXED_USE]
    Address:
      type: object
      required:
        - addressLine1
        - city
        - country
      properties:
        addressLine1:
          type: string
          example: "123 Main Street"
        addressLine2:
          type: string
          example: "Apt 4B"
        city:
          type: string
          example: "Lagos"
        state:
          type: string
          example: "Lagos State"
        country:
          type: string
          example: "Nigeria"
        postalCode:
          type: string
          example: "100001"
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tokenId:
          type: string
        name:
          type: string
          example: "Lagos Luxury Apartments"
        description:
          type: string
          example: "Premium residential complex in Victoria Island"
        address:
          $ref: '#/components/schemas/Address'
        valuation:
          type: number
          example: 5000000
        totalTokens:
          type: integer
          example: 50000
        availableTokens:
          type: integer
          example: 25000
        pricePerToken:
          type: number
          example: 100
        expectedYield:
          type: number
          example: 0.12
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, SOLD_OUT, INACTIVE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyCreation:
      type: object
      required:
        - name
        - address
        - valuation
        - totalTokens
        - pricePerToken
        - propertyType
      properties:
        name:
          type: string
          example: "Nairobi Office Complex"
        description:
          type: string
          example: "Modern office building in Westlands"
        address:
          $ref: '#/components/schemas/Address'
        valuation:
          type: number
          minimum: 1
          example: 3000000
        totalTokens:
          type: integer
          minimum: 1
          example: 30000
        pricePerToken:
          type: number
          minimum: 0.01
          example: 100
        expectedYield:
          type: number
          minimum: 0
          maximum: 1
          example: 0.10
        propertyType:
          $ref: '#/components/schemas/PropertyType'

    # Investment Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: object
    PropertiesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            properties:
              type: array
              items:
                $ref: '#/components/schemas/Property'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 20
                total:
                  type: integer
                  example: 100
                pages:
                  type: integer
                  example: 5
    PropertyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Property'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...